# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PDPS Backend API",
        "description": "API documentation for PDPS (Pradeshiya Sabha Development Project System) Backend",
        "contact": {
            "email": "admin@pdps.lk"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/sms/test": {
            "post": {
                "tags": [
                    "SMS Notifications"
                ],
                "summary": "Send test SMS",
                "description": "Send a test SMS message",
                "operationId": "635fdab60c119d7d2c52484034b09c41",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "+94771234567"
                                    },
                                    "message": {
                                        "type": "string",
                                        "maxLength": 160,
                                        "example": "Test SMS from PDPS system"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SMS sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "SMS sent successfully"
                                        },
                                        "message_sid": {
                                            "type": "string",
                                            "example": "SM1234567890abcdef"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/sms/payment-confirmation": {
            "post": {
                "tags": [
                    "SMS Notifications"
                ],
                "summary": "Send payment confirmation SMS",
                "description": "Send SMS confirmation for payment",
                "operationId": "600808f8c43a7087dfc72ac93ade5003",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "amount",
                                    "receipt_no",
                                    "service"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "+94771234567"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1500
                                    },
                                    "receipt_no": {
                                        "type": "string",
                                        "example": "RCP123456"
                                    },
                                    "service": {
                                        "type": "string",
                                        "example": "Water Bill Payment"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment confirmation SMS sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Payment confirmation SMS sent successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your API token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}